Migrate default WAF to v2
We want to add some other WAFs to this configuration, so start by updating
the existing one to the latest version.

This rule is used by Smokey to verify that a WAF is applied. My approach
will be to build more comprehensive WAFs for each public end point, and
include this rule in each of the new web ACLs. This will allow us to continue
to check that WAFs are applied.

The classic WAF's name had become a bit misleading after a bit, because although
it's called `CachePublicWebACL` it's now also applied to a bunch of other things that
are nothing to do with the cache LB.

You have to assign a capacity to `aws_wafv2_rule_group`s which is calculated
by adding up all the elements you use from [the rules statement list](https://docs.aws.amazon.com/waf/latest/developerguide/waf-rule-statements-list.html).
It's not super easy to figure out, but it allows users of your rule group
to ensure that their [rules come in under the maximum capacity](https://docs.aws.amazon.com/waf/latest/developerguide/how-aws-waf-works.html#aws-waf-capacity-units).

It's possible to modify rule groups, however the capacity is immutable because
consumers rely on it to be able to remain within limits. I've therefore set the
capacity at a bit more than it's using, so we can have some flexibility.

We're going to reuse the existing log stream, as it seems fit for purpose.
There's a [lambda that filters out logs](https://github.com/alphagov/govuk-aws/blob/main/terraform/lambda/WAFLogTrimmer/lambda.js)
that are generated by the `Default_Action`
(as these are requests that don't match any rule and will not be interesting).
This field is still good for [current WAF logs](https://docs.aws.amazon.com/waf/latest/developerguide/logging-fields.html),
so we'll continue to use this method.

I've only moved the public cache LB over to the new web ACL for now, but
will do the others soon.