Move other LBs to use the new WAF
In the previous commit we built a copy of the existing WAF using the newer
option in AWS - WAFv2.

In this commit, we migrate the remaining load balancers from the old version
to the new version and delete the old configuration.

The next step will be to build separate rule groups for the differing needs
of each load balancer and apply them individually. It's easier to do this
once we've moved everything over to the new version.

There is still a [Smokey check to check that a firewall](https://github.com/alphagov/smokey/blob/main/features/waf.feature)
using this rule exists and this still passes.