Set detailed exit status for terrform plans
Before there was a regex in Jenkins which interpreted `Plan ..`
as unstable builds. This is incorrect with newer versions of
Terraform which displays this line when doing applies with changes.

We now make terraform plan with detailed-exitcode so that it exits
with 2 when there are changes in the plan and Jenkins will mark the
build as unstable.