Output the IDs of public resources
These will be used in a separate project to better configure the web acls
for the public load balancers. The resources output are:

- IDs of public load balancers
- the ID of the existing kinesis data firehose

The load balancer IDs will be used in the new project to attach the new web
ACLs to the public load balancers.

The kinesis data firehose is used to [send the logs generated by the new WAFs to S3.](https://aws.amazon.com/premiumsupport/knowledge-center/waf-configure-comprehensive-logging/).
This happens at present with the default web ACL, but the plan is to move
that configuration out of the extremely large infra-public-services project
into something a bit more easily manageable. We'll reuse this kinesis data firehose
instance for the moment, until we need to do something different.

https://trello.com/c/IvZrVCL3/2869-setting-rate-limits-within-aws-waf