Add boilerplate for new WAFs project
This was inspired by the set up for [govuk-rds](https://github.com/alphagov/govuk-aws/tree/main/terraform/projects/app-govuk-rds)
which seems to be one of the few projects to uses a Terraform version > 1.

This project will allow us to remove some config from the infra-public-services
project (which is quite large), but should also more rapid deployment of
WAF changes should they be necessary.