Move shield and web ACL config to WAF project
This allows us to reduce the complexity of the infra-public-services project
and somewhat encapsulate the WAF configuration.

The config has been copied and pasted from one to the other and the syntax
updated to reflect the more recent TF version.

I also corrected it to be listed in alphabetical order based on the load
balancer name.