Datascience data: clean up script and remove tasks
We're removing any task, like the mongodb stuff.
that can be carried out by the application itself.
It makes it easier to make changes, as we don't have to run
terraform each time. Also it's better to have a cleaner
separation between platform setup (ie. this repo) and
application (the govuk-knowledge-graph repo)