Add VPC endpoint to route private subnets to S3
This allows us to communicate directly with S3 without using the NAT gateway.
It should save us a lot of money, as backups, crawler worker and lots of
other things send quite a lot of data that way.

For simplicity's sake I've added it to all private subnets, though they probably
don't all communicate with S3. By adding it to the module, it becomes default
that any new subnets will add it too.

This change:
- adds the S3 gateway endpoint to the VPC module and outputs the ID for use elsewhere
- modifies the private_subnet module to require the S3 gateway ID and add a route to it
- modifies infra-networking to pass the ID in when defining private subnets

Further reading:

- https://www.lastweekinaws.com/blog/the-aws-managed-nat-gateway-is-unpleasant-and-not-recommended/
- https://docs.aws.amazon.com/vpc/latest/privatelink/gateway-endpoints.html

https://trello.com/c/qi2ocydX/2651-use-vpc-endpoints-for-s3