Allow monitoring to access the mirror bucket
Smokey is currently failing because the monitoring machine is attempting to
use the new S3 gateway to retrive pages from the replica mirror and failing.

Looking at the policies on the bucket, we have locked them down to Fastly,
office and AWS NAT ip addresses, none of which will are used in this case.