Dev version: use a standard Ubuntu AMI and install neo4j manually
The community AMI used before has 2 drawbacks:

- it's complicated to update the neo4j version, we nee to rely on
  a third party to build a new AMI, or do it ourselves (and document it
  somewhere)

- we don't really know what's on the AMI. For instance awscli is installed,
  which we don't want (that's why we remove it at the beginning of userdata.tf)

So instead, we discard that AMI and use the standard ubuntu 22 one, and install
neo4j on it in the userdata script. This makes it explicit which version of
neo4j is installed, and how it was installed (here we just follow the neo4j
instructions).

We might find it better to eventually install neo4j in the
provision_knowledge_graph script in the govuk-knowledge-graph repo,
in order not have as little setup as possible in this shared repo.