Allocate EIPs for future use by GOV.UK Licensing
We've been trying for a long time to get Civica to update their ACLs to
allow our AWS environments to make API calls to their infrastructure, on
behalf of GOV.UK Licensing's local authority users.

We're starting to see some promising signs that they might finally do
this, and we want to make sure that we've got a reasonably future proof
set of IP addresses to give them.

Currently, API calls to Civica go through a VPN that runs to UK Cloud.

We'd like to remove this, and have API calls go through the NAT Gateways in our
production and staging VPCs. There are three NAT gateways in both environments,
so we have 6 IP addresses which need to be allowlisted.

In the future, there's a high chance we'll want to move Licensing into
its own VPC (or possibly its own AWS account). If this happens, we'll
end up with another set of IP addresses, and we'd have to do the whole
dance with Civica again.

To avoid this, this commit allocates 3 EIPs in our production VPC and 3
EIPs in our staging VPC. These will initially not be assigned to
anything, but we will ask Civica to add them to their allow lists. When
we eventually move Licensing, we'll be able to ask AWS support to move
these EIPs to the appropriate VPC / AWS account.