Staging and integration can read production backups
This grants access for the staging and integration DB admin machines to
read the production transition database backups. It is required for
their overnight env-sync tasks [1] [2].

The lack of this is preventing integration from syncing databases
overnight. Oddly it is working in staging where there seems to be an IAM
role policy added that isn't managed by Terraform.

[1]: https://github.com/alphagov/govuk-puppet/blob/950d713e663b081af763bff46d7db5c4cc4821db/hieradata_aws/class/integration/transition_db_admin.yaml#L2-L13
[2]: https://github.com/alphagov/govuk-puppet/blob/950d713e663b081af763bff46d7db5c4cc4821db/hieradata_aws/class/staging/transition_db_admin.yaml#L2-L13