Allow Fastly healthchecks to bypass rate limiting
Fastly healthchecks perform a HEAD request from each POP to each of the
IP addresses listed in our DNS record at a [frequency of our choosing](https://docs.fastly.com/en/guides/working-with-health-checks#creating-and-assigning-health-checks-using-vcl).
They occur more frequently than any other external request, so limit the
blanket rate limit we'd be able to apply here.

The header name is lower-cased when evaluating the rule.

We want them to succeed, so it seems fair to allow them.