Add a baseline rate limit rule for public cache
This rule is in count mode to start with.

This is not envisaged to trigger with normal usage patterns, so we'll tweak
the value to ensure that it's working as expected before settling on a final
value and flipping it to block mode.

We'll also have to consider adding custom responses for things that fail
this rule.