Add a log subscription filter
Cyber are migrating their ingestion lambda because of the deprecation (by AWS)
of something their old one relied on.

Here we add another filter pointing at their new lambda. The two will run
in parallel for a short time while we confirm that everything's working
and to ensure that we don't drop any logs.

https://trello.com/c/IJQFMwki/2984-update-cyber-splunk-endpoint