add explicit --lockfile=false flag to terraform-docs invocation
Builds were failing at the docs check linting stage, as the
version numbers generated in CI were using the version spec
given in the `required_providers` block, whereas running locally
they were using the exact version numbers in the .hcl lock file.

It looks like there was a bit of back-and-forth on this:
https://github.com/terraform-docs/terraform-docs/issues/517
but a flag --lockfile was introduced in terraform-docs v0.15:
https://github.com/terraform-docs/terraform-docs/releases/tag/v0.15.0

I assume the default setting for this differs in the Github Action
hosted runner compared to a .deb package install on Ubuntu, but
I have not been able to verify this.

This commit adds an explicit flag to the invocation so that whatever
environment you're running `tools/update-docs.sh` in, it will use
the same setting.