Add bucket for storing CSP reports
This creates a bucket that will be used to store CSP reports that will
be delivered by Kinesis Firehose. Firehose can emit both success data
and error data, both of these will be stored in this bucket under
different prefixes.

I doubt we'll want more than a short window into this data (it'll
probably be bombarded) so it only has a 30 day expiry.