Glue configuration for reading S3 data
This configures Glue to provide an Athena database to query CSP reports.
The columns are based on data that the CspReportsToFirehose lambda
produces, with partitions based on folder structure.

It expects ORC [1] as a file format for the files which is a columnar
data structure that is effective for Athena queries.

As we are expecting files to be in a partition directory structure, we
need a glue crawler to discover the partitions. This crawler has been
configured to not change the schema based on the data it finds and to
delete references to partitions that are missing.

I've set glue to run once an hour as we need it to run after a partition
is created but we're not clear on exactly when a partition may be
created (as CSP reports may be rather adhoc) thus this will mean the
data queried is perhaps delayed by an hour at times.

[1]: https://orc.apache.org/