Firehose configuration
We use Firehose as a means to buffer data from CSP reports into groups
written to S3, this means that a single S3 will likely contain multiple
CSP reports and be much more efficient to query than a single file per
report.

This firehose configuration expects a JSON put and it will output the
data into S3 after converting it to ORC file format.

The conversion to the ORC format was decided because using JSON produced
files that Athena could not query, as each JSON record didn't have a
separating new line. When confronted with this I decided that it was
better to convert the data into a format well suited for querying,
rather than say writing a lambda that adjusted the JSON.

I chose the ORC format over Paraquet as I found it easier to install
tools to view the files (installing parquet-tools is rather confusing as
brew has deprecated the package and I can't tell if it's actually
maintained), and then when I did get tools parquet didn't seem to like
the parquet format AWS uses with various warnings. Instead, using ORC was
quite straight forward and orc-tools was easy to install and use.

A pain point I found is that to create a Kinesis Delivery Stream I
needed to provide the Kinesis role with access to read Glue tables,
however the Kinesis role also needed permissions to run the created
Kinesis Delivery Stream. To resolve this cyclic dependency I split the
permissions into different resources.