Configure API Gateway
This sets up an API Gateway HTTP API as a proxy for the Lambda installed
in the previous commit.

This has a just single endpoint, POST /report, which integrates with the
aforementioned Lambda. As that Lambda has been configured for the 2.0
[1] payload this API gateway is configured to use that.

In order to call the Lambda we need to add a permission to the lambda
that permits invoking from the api-gateway.

I did wonder about configuring CORS headers for this - first as a means
to restrict the origins, second as a way to ensure success. However on
investigation I can't find any usage of CORS for browsers with
report-uri, so I believe browsers don't use it. It looks like it would
be needed should we switch to the upcoming Reporting API [2].

[1]: https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html#http-api-develop-integrations-lambda.proxy-format
[2]: https://developer.chrome.com/articles/reporting-api/#option-2-build-and-operate-your-own-report-collector