Add custom domain handling
This configures domain names for the API gateway they accept
configuration of a domain name and a certifcate ARN to specify the
domain and the certificate.

The certificate must cover both the specified domain name and
csp-reporter.<environment>.govuk.digital. The domain name, if it is not
csp-reporter.<environment>.govuk.digital must be set as a CNAME to that.

It's also mostly Roch's work that I've copied and pasted!