Create a bucket for Athena Query results
This creates an S3 bucket that is intended for the storage of Athena query
results conducted via the AWS web UI. It is has been done to reduce the
frustration of developers taking an adhoc approach and setting a new
bucket each time they run an Athena query.

This bucket is to be configured as the query result destination for the
default workgroup, primary. This bucket will have to be connected to that
workgroup via clickops as I can't find a good way to edit that automatically
created workgroup, and it seems low value in using terraform to create a new
workgroup that isn't the default (as I expect every user would forget to
change it).

The clickops process is to go to Athena, select workgroups from the
right bar (under administration). Then go into Query result
configuration and select this bucket as the Location of query results.