Bump Terraform version
The `plan`/`apply` worked in Integration and Staging, but failed
on Production with:

```
Error:
Terraform doesn't allow running any operations against a state
that was written by a future Terraform version. The state is
reporting it is written by Terraform '0.12.30'

A newer version of Terraform is required to make changes to the current
workspace.
```

Trello: https://trello.com/c/LVGgeLG7