Remove UK Cloud VPN
This is no longer required because we've worked with Civica to allowlist the
IPs of our AWS NAT Gateways for the vast majority of their local authority
customers.

Additionally, UK Cloud is in liquidation and is being shut down. The
liquidators only expect the platform to be operational until 10 February 2023,
and in the meantime the liquidators have been increasing prices exponentially
to cover the costs of operating the platform with an ever diminished customer
base.

I have notified UK Cloud and the liquidators that from the end of December 2022
we no longer require our account.