Upgrade NodeJS versions used in infra-mirror-bucket lambda
This was first attempted in #1652 but had to be reverted in #1688,
as the Node version was incompatible with the version of
terraform-provider-aws we were using, leading to this error on a
`plan`:

```
Error: aws_lambda_function.url_rewrite: expected runtime to be one of [dotnetcore1.0 dotnetcore2.0 dotnetcore2.1 go1.x java8 java11 nodejs4.3 nodejs4.3-edge nodejs6.10 nodejs8.10 nodejs10.x nodejs12.x provided python2.7 python3.6 python3.7 python3.8 ruby2.5], got nodejs16.x
```

[v3.75.2](https://github.com/hashicorp/terraform-provider-aws/releases/tag/v3.75.2)
introduces the `nodejs16.x` runtime to Lambdas, so we bump to at
least version 3.76 to ensure the runtime is available.
This required updating Terraform from v0.11.15 to v0.13.6 (via
v0.12.30) in #1691 and #1692.

Trello: https://trello.com/c/PMsAXSoT/3029-review-govuks-usage-of-node-2