WIP: Upgrade NodeJS versions used in our AWS Lambdas
This was first attempted in #1652 but had to be reverted in #1688,
as the Node version was incompatible with the version of
terraform-provider-aws we were using, leading to this error on a
`plan`:

```
Error: aws_lambda_function.url_rewrite: expected runtime to be one of [dotnetcore1.0 dotnetcore2.0 dotnetcore2.1 go1.x java8 java11 nodejs4.3 nodejs4.3-edge nodejs6.10 nodejs8.10 nodejs10.x nodejs12.x provided python2.7 python3.6 python3.7 python3.8 ruby2.5], got nodejs16.x
```

[v3.75.2](https://github.com/hashicorp/terraform-provider-aws/releases/tag/v3.75.2)
introduces the `nodejs16.x` runtime to Lambdas, so we bump to at
least version 3.76 to ensure the runtime is available.

Also updates the Terraform version to at least 1.1, as all the
other examples of specifying AWS provider seem to have at least
that version, and those changes were otherwise having seemingly
no effect.

This, in turn, needed an explicit `.terraform-version` file, as
otherwise the `plan` would fail with:

```
Error: The currently running version of Terraform doesn't meet the
version requirements explicitly specified by the configuration.
Please use the required version or update the configuration.
Note that version requirements are usually set for a reason, so
we recommend verifying with whoever set the version requirements
prior to making any manual changes.

  Module: root
  Required version: ~> 1.1
  Current version: 0.11.15
```

Trello: https://trello.com/c/PMsAXSoT/3029-review-govuks-usage-of-node-2