Resolve lint warnings
Resolves warnings like:

```
Warning: Quoted type constraints are deprecated

  on main.tf line 12, in variable "aws_region":
  12:   type        = "string"

Terraform 0.11 and earlier required type constraints to be given in quotes,
but that form is now deprecated and will be removed in a future version of
Terraform. To silence this warning, remove the quotes around "string".

(and 18 more similar warnings elsewhere)

Warning: Interpolation-only expressions are deprecated

  on main.tf line 130, in provider "aws":
 130:   region  = "${var.aws_region}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.

(and 80 more similar warnings elsewhere)

Warning: Quoted references are deprecated

  on main.tf line 248, in resource "aws_s3_bucket" "govuk-mirror-replica":
 248:   provider = "aws.aws_replica"

In this context, references are expected literally rather than in quotes.
Terraform 0.11 and earlier required quotes, but quoted references are now
deprecated and will be removed in a future version of Terraform. Remove the
quotes surrounding this reference to silence this warning.
```