Make it clearer that `cloudfront_create` is a boolean
We only define it in one place in govuk-aws-data, and its value
there is `true`:
https://github.com/alphagov/govuk-aws-data/blob/310595a7ab51a859f9b9074ab5447e19d3f391bb/data/infra-mirror-bucket/production/common.tfvars#L6

Whilst this is unlikely to fix the `plan` error below, it should
at least make it easier for future developers to reason about why
it is happening.

```
Error: Missing resource instance key

  on main.tf line 499, in resource "aws_cloudfront_distribution" "www_distribution":
 499:     acm_certificate_arn      = data.aws_acm_certificate.www.arn

Because data.aws_acm_certificate.www has "count" set, its attributes must be
accessed on specific instances.

For example, to correlate with indices of a referring resource, use:
    data.aws_acm_certificate.www[count.index]
```