Fix "Missing resource instance key" error
We use 'count' on each 'aws_cloudfront_distribution' as a way
of initialising or not initialising the resource, based on the
boolean value 'cloudfront_create'.

In other words, the count is only ever 0 or 1, so this code should
only ever be being applied when the count is 1, in which case the
index of the resource is always 0.