Temporarily bump NodeJS runtime to v12
The `apply` is failing because the version of Node we're on is no
longer supported:

```
Error: Error modifying Lambda Function Configuration url_rewrite: InvalidParameterValueException: The runtime parameter of nodejs10.x is no longer supported for creating or updating AWS Lambda functions. We recommend you use the new runtime (nodejs16.x) while creating or updating functions.
{
  Message_: "The runtime parameter of nodejs10.x is no longer supported for creating or updating AWS Lambda functions. We recommend you use the new runtime (nodejs16.x) while creating or updating functions.",
  Type: "User"
}

  on main.tf line 618, in resource "aws_lambda_function" "url_rewrite":
 618: resource "aws_lambda_function" "url_rewrite" {
```

This is an ironic thing to fail on, as the entire point of us
upgrading the Terraform is to upgrade the Node version, but we're
unable to jump straight to the v1.1 version of Terraform required
to support Node v16. So this is an attempt to fix the immediate
error we're seeing, even if we're planning to bump to v16 very
shortly.