Upgrade infra-public-services Terraform from 0.11.15 to 0.12.30
The long term aim is to upgrade to at least 0.13 so that we can
bump the NodeJS runtime for the lambda in this project, from v12
to v16.

However, based on a similar upgrade in #1691, it is not possible
to upgrade straight to 0.13, as that would be considered two
major versions ahead of 0.11. The `plan` would fail. So this
0.12 upgrade is a necessary stopgap.