Fixup: fix "Unknown block type" for /modules/aws folder
```
Error: Unsupported block type

  on ../../modules/aws/lb/main.tf line 367, in resource "aws_cloudwatch_metric_alarm" "elb_httpcode_elb_5xx_count":
 367:   dimensions {

Blocks of type "dimensions" are not expected here. Did you mean to define
argument "dimensions"? If so, use the equals sign to assign it a value.
```

and

```
Error: Unsupported block type

  on ../../modules/aws/alarms/elb/main.tf line 180, in resource "aws_cloudwatch_metric_alarm" "elb_surgequeuelength":
 180:   dimensions {

Blocks of type "dimensions" are not expected here. Did you mean to define
argument "dimensions"? If so, use the equals sign to assign it a value.
```