Update `subnets` list syntax to work with Terraform v0.12
We were seeing errors like:

```
Error: Incorrect attribute value type

  on main.tf line 1496, in resource "aws_elb" "jumpbox_public_elb":
1496:   subnets         = ["${data.terraform_remote_state.infra_networking.outputs.public_subnet_ids}"]
    |----------------
    | data.terraform_remote_state.infra_networking.outputs.public_subnet_ids is list of string with 3 elements

Inappropriate value for attribute "subnets": element 0: string required.
```

This was causing cascading failures elsewhere in the `plan`, where
modules were not expecting to be passed an array:

```
Error: Incorrect attribute value type

  on ../../modules/aws/lb/main.tf line 212, in resource "aws_lb" "lb":
 212:   subnets            = "${var.subnets}"
    |----------------
    | var.subnets is list of list of string with 1 element

Inappropriate value for attribute "subnets": incorrect set element type:
string required.
```

It appears that in Terraform 0.11, you could pass a list inside
quotes and it would be treated as a list, but in 0.12 it is now
treated as a string, so we need to do some reformatting for
consistency and to ensure it is properly treated as a list
everywhere.

However, making this change alone caused the terraform fmt GitHub
Action linter to fail, because _not_ using quotes is unrecognised
in Terraform 0.11. So we've had to bump the Terraform version for
the parent project too.