Add a monitor for non-GET requests
We might want to add different rate limits for different request types. This
will allow us to tweak limits for certain HTTP methods based on a regex.

This is in count only mode at present.

We start by defining the regex of the pattern we want to match, as a
`aws_wafv2_regex_pattern_set` called `method_allow_list_pattern`.

We then reference it in the rate limit rule by using a `regex_pattern_set_reference_statement`.
It would be more readable to use a `regex_match_statement`, but apparently
it's not possible to embed them in this type of rule.

We want to create a sort of allow list, so that lightweight requests are
subject to the more lenient rate limit that occurs before, but are not inhibited
by this one. Because of that, the regex we define, is for the "good" methods.

We then `not` the match status for that regex, so that only things that don't
match our regex are subject to this rate limit.

We don't want to perform any text transformation on this comparison, as we
want to catch anything weird in our net of things to be cautious about.
Standard HTTP methods are upper case, and so should the reasonable requests
to our system be.