Reset the priority numbers
I want to make the priority more overt. There's no reason to leave a gap
if we control all the options (which we do).

Also, if any edits are made in the UI (such as tweaking counts, or trialling
new rules) AWS appears to do this implicitly. That means the next Terraform
run generates a diff that you weren't quite expecting.