Configure AmazonMQ via Lambda functions
As we deploy Terraform code from Jenkins in integration regardless
of which environment it is targeting, this means the Jenkins instance
has access to the AWS APIs but not to any resources in any other
environment.

The way around this is to create Lambda functions in the target
environments, which can post the JSON config when invoked by Jenkins.
This has been configured successfully for the Publishing AmazonMQ.

For more detail on the prior art see https://github.com/alphagov/govuk-aws/pull/1701