Allow TF overrides to be defined in govuk-aws-data
Updates the `build-terraform-project` script to apply TF override files defined in the corresponding project directory in `govuk-aws-data`.

We symlink the override files instead of copying them, so that we can clean them up after running the TF command while preserving any override files that may be present in _this_ repo.