Add dependabot.yml
By default Dependabot is raising PRs to update the pip dependencies for our lambdas. These lambdas are not tested as part of CI, and deploying them involves an additional build step which is not automated. For these reasons, I don't think it's appropriate for Dependabot to raise PRs for them.

This commit adds a Dependabot config that excludes pip.