Remove old IAM policy from infra-mirror-bucket
The mirrorer has been moved to Kubernetes, and the old mirrorer app has been deleted from AWS. This is causing Terraform plans/appies for infra-mirror-bucket to fail, because it defines an IAM policy to allow the (now-deleted) mirrorer app to write to the mirror bucket.

This commit removes the redundant IAM policy.

Terraform plan for this change is a no-op for all 3 environments.