Add binding for `search-api-v2` bulk sync
Like the existing functionality in `search-api`, we want to add the
ability to enqueue a "bulk reindex/sync" for `search-api-v2`. This
deliberately listens to a different routing key from `search-api` to
avoid pointlessly reindexing documents there as part of any v2 work, and
uses the same queue as regular indexing for simplicity.