Enable timelock for production DB backups.
Compliance mode = nobody can delete[1] until the time expires, not even
root.

Also set the no-public-access guardrails while we're at it.

[1] Strictly speaking, it's at least theoretically possible to delete
timelocked stuff by persuading Amazon to delete the entire account that
contains the buckets, but that's probably not a very realistic attack
scenario. (If somebody thinks it is, then we need to be multi-cloud
already.)