Remove no-longer-useful linters, scripts etc.
Keep the terraform fmt check (still helps fight entropy) and the YAML
lint (not sure it's doing much any more but it's fast and self-contained
so no reason to delete it).

Ditch the Ruby stuff because it's no longer checking anything important
and it generates maintenance overhead with all the dependency updates.

Especially ditch the terraform-docs check - it's of negative value at
this point.