Log all requests through backend_public WAF ACL
Rationale: We recently added support for JA3 denylisting on the
backend_public ACL, but the only place we log JA3 fingerprints is in the
WAF logs. Currently the backend_public ACL is configured to only log
requests that were blocked or counted, so we have no way to find the JA3
fingerprints of traffic that was allowed, but that we want to add to the
JA3 denylist.

This will increase the number of requests being logged by this ACL, but
the amount is around the same order of magnitude as the volume of
requests we already block (and therefore log) in the cache_public ACL.
The `aws-waf-logs-backend-public-production` log group only has 1 month
of retention.