Remove references to sg_management_id
This was the security group that let us SSH into instances in the
pre-replatforming world.

It no longer exists, but its ID is still in the now deleted
infra-security-groups project's statefile.

None of these projects have been applied since the security group was
deleted, but clearly they've all been removed from the security group
(since it no longer exists). Or in the case of related links, they're
probably just not applied at all.

This commit at least fixes an obviously broken bit of config, making it
slightly more likely that we'd be able to apply these projects if we
wanted to.

It's enough to fix app-elasticsearch6, which is the one I care about
right now.